+++ neo/Implementations/Wallets/peewee/test_user_wallet.py
@@ -150,7 +150,15 @@ class UserWalletTestCase(WalletFixtureTestCase):
 
         wallet.AddNEP5Token(token)
 
+        wallet = self.GetWallet1(recreate=True)  # re-open wallet
+
         self.assertEqual(len(wallet.GetTokens()), 1)
+
+        wallet.DeleteNEP5Token(token.ScriptHash)
+
+        wallet = self.GetWallet1(recreate=True)  # re-open wallet
+
+        self.assertEqual(len(wallet.GetTokens()), 0)
 
     def test_8_named_addr(self):